+++ examples/tokio-tcp-stream-console-logger.rs
@@ -1,7 +1,7 @@
 use logged_stream::ConsoleLogger;
 use logged_stream::DefaultFilter;
-use logged_stream::HexDecimalFormatter;
 use logged_stream::LoggedStream;
+use logged_stream::LowercaseHexadecimalFormatter;
 use std::env;
 use tokio::io::AsyncReadExt;
 use tokio::io::AsyncWriteExt;
@@ -42,7 +42,7 @@ async fn main() {
 
     let mut client = LoggedStream::new(
         net::TcpStream::connect("127.0.0.1:8080").await.unwrap(),
-        HexDecimalFormatter::new(None),
+        LowercaseHexadecimalFormatter::new_default(),
         DefaultFilter::default(),
         ConsoleLogger::new_unchecked("debug"),
     );